requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.5.0
gpxpy==1.6.1
python-dotenv==1.0.0
python-dotenv==1.0.0
requests==2.31.0
plotly-5.18.0
pandas==2.1.3
